refactor(#53): baseMoney 타입 변경
int는 타입 표현 불가 -> Long으로 변경

+++ src/main/kotlin/team/mozu/dsm/domain/lesson/model/Lesson.kt
@@ -11,7 +11,7 @@ data class Lesson(
     val lessonName: String,
     val maxInvRound: Int,
     val curInvRound: Int,
-    val baseMoney: Int,
+    val baseMoney: Long,
     val lessonNum: String,
     val isStarred: Boolean,
     val isDeleted: Boolean,
